Fixed zoom and drag handler
- Unified them into one gesture handler to prevent conflicts
